+++ projects/nb-form/src/lib/constants/control-err-type.enum.ts
@@ -1,4 +1,4 @@
-export enum NbControlErrTypeEnum {
+export enum NbControlErrType {
   REQUIRED = 'required',
   FILE_MAX_SIZE = 'fileMaxSize',
   FILE_MIN_SIZE = 'fileMinSize',
@@ -10,3 +10,8 @@ export enum NbControlErrTypeEnum {
   ARR_MIN_LENGTH = 'arrMinLength',
   WHITESPACE = 'whitespace',
 }
+
+/**
+ * @deprecated use 'NbControlErrType' please
+ */
+export const NbControlErrTypeEnum = NbControlErrType;