+++ projects/nb-form/src/lib/validators/equal.ts
@@ -1,7 +1,7 @@
 
 import { AbstractControl } from '@angular/forms';
 import { isEqual } from 'lodash-es';
-import { NbControlErrTypeEnum } from '../constants';
+import { NbControlErrType } from '../constants';
 
 export const equal = (compared: AbstractControl, immediately: boolean = true) => {
     return (target: AbstractControl) => {
@@ -10,6 +10,6 @@ export const equal = (compared: AbstractControl, immediately: boolean = true) =>
         }
         return isEqual(target.value, compared.value)
             ? null
-            : { [NbControlErrTypeEnum.EQUAL]: true };
+            : { [NbControlErrType.EQUAL]: true };
     }
 }