+++ projects/nb-form/src/lib/validators/file-size.ts
@@ -1,5 +1,5 @@
 import { AbstractControl, ValidatorFn } from '@angular/forms';
-import { NbControlErrTypeEnum } from '../constants';
+import { NbControlErrType } from '../constants';
 
 export const fileSize = (fileSize: { max?: number; min?: number }): ValidatorFn => {
   return (control: AbstractControl) => {
@@ -10,11 +10,11 @@ export const fileSize = (fileSize: { max?: number; min?: number }): ValidatorFn 
     }
 
     if (fileSize.max && file.size > fileSize.max) {
-      return { [NbControlErrTypeEnum.FILE_MAX_SIZE]: true };
+      return { [NbControlErrType.FILE_MAX_SIZE]: true };
     }
 
     if (fileSize.min && file.size < fileSize.min) {
-      return { [NbControlErrTypeEnum.FILE_MIN_SIZE]: true };
+      return { [NbControlErrType.FILE_MIN_SIZE]: true };
     }
 
     return null;