+++ src/app/app.module.ts
@@ -1,19 +1,19 @@
 import { NgModule } from '@angular/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { BrowserModule } from '@angular/platform-browser';
-import { NbTransLangEnum, NbTransModule, NbTransService, NB_TRANS_LOADER } from '@bigbear713/nb-trans';
+import { NbTransLang, NbTransModule, NbTransService, NB_TRANS_LOADER } from '@bigbear713/nb-trans';
 import {
   NbFormModule,
-  NbControlErrTypeEnum,
-  NB_CONTROL_COMMON_ERR_INFO_TOKEN
+  NbControlErrType,
+  NB_CONTROL_COMMON_ERR_INFO,
 } from 'nb-form';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { HttpClient, HttpClientModule } from '@angular/common/http';
 @NgModule({
-  declarations: [
-    AppComponent
-  ],
+  declarations: [	
+    AppComponent,
+   ],
   imports: [
     BrowserModule,
     HttpClientModule,
@@ -25,11 +25,11 @@ import { HttpClient, HttpClientModule } from '@angular/common/http';
   ],
   providers: [
     {
-      provide: NB_CONTROL_COMMON_ERR_INFO_TOKEN,
+      provide: NB_CONTROL_COMMON_ERR_INFO,
       useFactory: (transService: NbTransService) => ({
-        [NbControlErrTypeEnum.FILE_TYPE]: transService.translationAsync('errors.fileType'),
-        [NbControlErrTypeEnum.FILE_MIN_SIZE]: 'The file min size is 100KB!',
-        [NbControlErrTypeEnum.EQUAL]: '两个控件值不相等',
+        [NbControlErrType.FILE_TYPE]: transService.translationAsync('errors.fileType'),
+        [NbControlErrType.FILE_MIN_SIZE]: 'The file min size is 100KB!!!!',
+        [NbControlErrType.EQUAL]: '两个控件值不相等',
       }),
       deps: [NbTransService]
     },
@@ -37,8 +37,8 @@ import { HttpClient, HttpClientModule } from '@angular/common/http';
       provide: NB_TRANS_LOADER,
       useFactory: (http: HttpClient) => {
         return {
-          [NbTransLangEnum.ZH_CN]: () => http.get('./assets/localization/zh-CN/translations.json'),
-          [NbTransLangEnum.EN]: () => http.get('./assets/localization/en/translations.json'),
+          [NbTransLang.ZH_CN]: () => http.get('./assets/localization/zh-CN/translations.json'),
+          [NbTransLang.EN]: () => http.get('./assets/localization/en/translations.json'),
         };
       },
       deps: [HttpClient]